Working Stable

+++ app/Http/Livewire/Form.php
@@ -12,33 +12,71 @@ class Form extends Component
 		return view('livewire.form');
 	}
 
-	public $name, $description, $categorie, $amount, $price, $type;
+	public $ID, $name, $description, $categorie, $amount, $price, $type;
+	protected $listeners = [
+		'editArticle' => 'editArticle',
+		'deleteArticle' => 'deleteArticle',
+		'confirmDeleteArticle' => 'confirmDeleteArticle',
+	];
+
+	protected $rules = [
+		'name' => 'required',
+		'description' => 'required',
+		'categorie' => 'required',
+		'amount' => 'required',
+		'price' => 'required'
+	];
 
 	public function unloadArticle() {
 		$this->name = "";
 		$this->description = "";
 		$this->categorie = "";
-		$this->amount = 0;
-		$this->price = 0.0;
-	}
-
-	public function addArticle() {
-		Articles::create([
-			'name' => $this->name,
-			'description' => $this->description,
-			'categorie' => $this->categorie,
-			'amount' => $this->amount,
-			'price' => $this->price
-		]);
-
-		$this->resetForm();
-	}
-
-	public function deleteArticle($id) {
-		Articles::findOrFail($id)->delete();
+		$this->amount = "";
+		$this->price = "";
+		$this->type = 0;
 	}
 
 	public function editArticle($id) {
+
+		$a = Articles::findOrFail($id);
+		if(isset($a)) {
+			$this->ID = $a->id;
+			$this->name = $a->name;
+			$this->description = $a->description;
+			$this->categorie = $a->categorie;
+			$this->amount = $a->amount;
+			$this->price = $a->price;
+			$this->type = 1;
+		}
+	}
+
+	public function addArticle() {
+		$validate = $this->validate($this->rules);
+		if($this->type == 0) {
+			Articles::create($validate);
+		} else if($this->type == 1) {
+			Articles::findOrFail($this->ID)->update($validate);
+		} 
+		$this->emitTo('table','updateTable');	
+		$this->unloadArticle();
+	}
+
+	public function confirmDeleteArticle($id) {
+		$this->dispatchBrowserEvent('confirm', [
+			'name' => Articles::find($id)->name,
+			'id' => $id
+		]);
+	}
+
+	public function deleteArticle($op, $id) {
+		if($op) {
+			Articles::findOrFail($id)->delete();
+			$this->emitTo('table','updateTable');	
+		}
+	}
+
+
+	public function updateArticle($id) {
 		Articles::findOrFail($id)->update([
 			'name' => $this->name,
 			'description' => $this->description,
@@ -46,6 +84,7 @@ class Form extends Component
 			'amount' => $this->amount,
 			'price' => $this->price
 		]);
-	}
 
+		$this->unloadArticle();
+	}
 }
